test: Run test for supabase_flutter on web as well (#1140)
* test: support web test for supabase_flutter

* ci: add chrome test

* ci: run chrome test only on latest version

* ci: add wasm test

* ci: correct matrix check

* refactor: fix analytics

* refactor: migrate web localstorage to web package
