+++ .github/workflows/supabase_flutter.yml
@@ -60,15 +60,23 @@ jobs:
           melos bootstrap
 
       - name: flutterfmt
-        if: ${{ matrix.sdk == '3.x'}}
+        if: ${{ matrix.flutter-version == '3.x'}}
         run: dart format lib test -l 80 --set-exit-if-changed
 
       - name: analyzer
-        if: ${{ matrix.sdk == '3.x'}}
+        if: ${{ matrix.flutter-version == '3.x'}}
         run: flutter analyze --fatal-warnings --fatal-infos .
 
       - name: Run tests
         run: flutter test --concurrency=1
+
+      - name: Run tests on chrome with js
+        if: ${{ matrix.flutter-version == '3.x'}}
+        run: flutter test --platform chrome
+
+      - name: Run tests on chrome with wasm
+        if: ${{ matrix.flutter-version == '3.x'}}
+        run: flutter test --platform chrome --wasm
 
       - name: Run tests with downgraded app_links
         run: |
