+++ packages/supabase_flutter/lib/src/local_storage_web.dart
@@ -1,17 +1,16 @@
-// ignore: avoid_web_libraries_in_flutter
-import 'dart:html' as html;
+import 'package:web/web.dart';
 
-final _localStorage = html.window.localStorage;
+final _localStorage = window.localStorage;
 
 Future<bool> hasAccessToken(String persistSessionKey) async =>
-    _localStorage.containsKey(persistSessionKey);
+    _localStorage.getItem(persistSessionKey) != null;
 
 Future<String?> accessToken(String persistSessionKey) async =>
-    _localStorage[persistSessionKey];
+    _localStorage.getItem(persistSessionKey);
 
 Future<void> removePersistedSession(String persistSessionKey) async =>
-    _localStorage.remove(persistSessionKey);
+    _localStorage.removeItem(persistSessionKey);
 
 Future<void> persistSession(
         String persistSessionKey, persistSessionString) async =>
-    _localStorage[persistSessionKey] = persistSessionString;
+    _localStorage.setItem(persistSessionKey, persistSessionString);
