+++ packages/supabase_flutter/pubspec.yaml
@@ -22,6 +22,7 @@ dependencies:
   path_provider: ^2.0.0
   shared_preferences: ^2.0.0
   logging: ^1.2.0
+  web: '>=0.5.0 <2.0.0'
 
 dev_dependencies:
   dart_jsonwebtoken: ^2.4.1
